Rename OneSignalCacheCleaner to OSNotificationDataController
* OSNotificationDataController will be the new class responsible for handling notification table access
